Update to internal 0.59

+++ meta-openbmc-mods/meta-common/recipes-phosphor/watchdog/system-watchdog/watchdog-reset.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 
+echo "Watchdog being started by $1" > /dev/kmsg
+
 if /sbin/fw_printenv bootfailures -n | grep -q 3; then
     exit 0 # passed boot limit, user started again on purpose
 fi
 
-echo "Watchdog Failure Limit Reached, Failed Processes:" > /dev/kmsg
-systemctl --failed --no-pager | grep failed > /dev/kmsg
+if test -f "/tmp/nowatchdog"; then
+    echo "Not resetting due to nowatchdog file" > /dev/kmsg
+    exit 0
+fi
+
 echo "Log as follows:" > /dev/kmsg
 journalctl -r -n 100 | while read line; do echo $line > /dev/kmsg; done
